Removes issue tracker style
When using GitHub, PR’s are already implicitly connected
to branches. 

We should prefer to open a PR *early* and *often*
with the expressed goal of early feedback.

Make liberal use of the `WIP` prefix and the `Close`
PR button. It’s ok to close PR’s, it’s just git